+++ LSM303DLH/examples/Serial/Serial.pde
@@ -14,19 +14,19 @@ void loop() {
 
   Serial.print("A ");
   Serial.print("X: ");
-  Serial.print(compass.a.x);
+  Serial.print((int)compass.a.x);
   Serial.print(" Y: ");
-  Serial.print(compass.a.y);
+  Serial.print((int)compass.a.y);
   Serial.print(" Z: ");
-  Serial.print(compass.a.z);
+  Serial.print((int)compass.a.z);
 
   Serial.print(" M ");  
   Serial.print("X: ");
-  Serial.print(compass.m.x);
+  Serial.print((int)compass.m.x);
   Serial.print(" Y: ");
-  Serial.print(compass.m.y);
+  Serial.print((int)compass.m.y);
   Serial.print(" Z: ");
-  Serial.println(compass.m.z);
+  Serial.println((int)compass.m.z);
   
   delay(100);
 }
